Fix capture of memory maps resulting in an empty file
